fix(gui): レビュー ①② — gshell 終了時に gui_handler/pump を解除、lease_mask を 0xFF7E に (14 色)
① kernel/gui.c gui_owner_exit: owner==GUI_SHELL_OWNER なら g_gui_handler/g_gui_pump を
   クリア。gshell 退去後に shell.bin が同じ 0x300000 に載っても旧ポインタを呼ばない (K2 対策)。
② gfx/backend_pc98.c lease_mask 0x7F7E→0xFF7E (index 1-6・8-15 の 14 色、契約 G8。0x7F7E は
   index15 も禁止する誤り)。TASK_H1 の転記ミスも訂正。
+ gdi_test の line() スカラ座標キャストを i32 統一に追従。

+++ userland/rust/gdi_test/src/lib.rs
@@ -321,10 +321,10 @@ fn draw_clip_demo(screen: SurfaceId) {
     /* はみ出す対角線 (クリップされる)。 */
     g::line(
         screen,
-        (cx - 30) as i16,
-        (cy - 30) as i16,
-        (cx + cw + 30) as i16,
-        (cy + ch + 30) as i16,
+        cx - 30,
+        cy - 30,
+        cx + cw + 30,
+        cy + ch + 30,
         Style::pen(GUI_COLOR_TEXT),
     );
     /* 内側クリップを入れ子に。別の色で塗る。 */
